Move instance splitter
