--- src/gluonts/torch/model/tft/transformation.py
@@ -1,1 +1,0 @@
-from gluonts.mx.model.tft._transform import TFTInstanceSplitter
